pyproject.toml
[project]
name = "explainable-ai-for-news-integrity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.4.0",
    "datasets>=2.14.0",
    "google-generativeai>=0.3.0",
    "groq>=0.37.1",
    "nltk>=3.9.2",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pyspark>=3.5.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "streamlit>=1.28.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
]
